swap names for host files
